+++ src/test/java/nextstep/ladder/domain/RandomLadderGeneratorTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class RandomMyLadderGeneratorTest {
+class RandomLadderGeneratorTest {
 
     private LadderGenerator generator;
